WIP started man page of rainsd.
Just a first draft of the rainsd man page which has some of the options
discovered so far with an attempt at defining what they do. Need to
finish documenting the options then see how to present the configuration
format in the man page / add more command line options in addition to the
config file?

Next steps are probably to spin up some example servers and see how to
configure them / query them. This would give a better idea of how the
binary should be worked with and what interface the client side should
have.