Initial commit for integration tests.
Add a new binary which will run the integration tests on the RAINS system.
This binary generates a one-time-use configuration for the whole system
in a temporary directory which contains a root server and subdomains.

This commit includes the confuration and key generation for the root server.
Future commits will contain:
* Config generation for level 2 domains (e.g. ccTLD's),
* A system for starting the servers from the test framework,
* Zone file generation
* Rainspub based zone publication in the test framework,
* Rainsdig based querying in the test framework,

The aim of this is to be able to automate end-to-end testing of the system.