Integration test supports publishing root zone.
The integration testing tool now supports initializing all nameservers,
generating the root zone file, and running rainspub to push the zones
to the root nameserver.

Commands to reproduce:

make
./build/integration --alsologtostderr
... some output explaining what is going on ...

`integration` will exit with an error if something goes wrong.

The next stage is to push the A records to the level2 nameservers.
After that, rainsdig will be used to query the servers and ensure
the right response is returned.