Integration tests now support publishing to L2TLD.
The l2TLD's are now populated with some dummy domains that are generated
from a list passed in via an argument to the integration test program.

The next step is to verify the querying of these records works. Before
this can happen rainsdig has to be debugged to see why it hangs when
passing in a name to query for.