First integration of SCION into rainsd.
This commit adds in support for listening via SCION in rainsd.

To get this to compile properly, govendor needs to be used. This is because the
scion repository has specific version requirements on libraries such as
quic-go.

The steps for using govendor are as follows:

govendor init
govendor add +e
govendor fetch +m

These commands must be performed from the rains folder (i.e. the folder with .git).