Update and clean up SCION connection initialisation  (#231)
Preparing for the next update to scionproto/scion master.

Add wrapper layer for initialising snet, creating connections etc. As this has almost identical functionality with the same wrapper that is used in scion-apps, this repo now contains a stripped-down copy of the appnet library. This is far from optimal but stripping out the common functionality into a shared library seems to add many complications for little gain -- the shared library could not include the hostnames functionality (as this depends on RAINS) so yet another wrapper would be required.
Previously, similar functionality was just scattered across the code base so the problem was here anyway, just less obvious. We'll fix this if it starts to hurt.

Try to centralise many of the previously scattered type checks for SCION vs TCP transport. These checks were still compiling correctly but were not working any more, due to changed types in snet (hurrah for type safety...).

Try to sanitize some of the most egregious constructs encountered during this process, accepting the risk of breaking things. The integration tests are passing, so...

* Adapt CI setup to new scion master, split into two stages

Split the CI into two jobs, one for only build and checks of RAINS,
one for integration tests with running SCION stack.

Newer versions of SCION don't have fewer messy python dependencies so
installing them in circleci/golang container actually works.
Also, no need to chache the python deps because there are so few left.
This makes the setup a bit easier, I think.

* Checkout SCION for integration tests based on version specified in go.mod