Combine ScionIPv4/v6 address types, binary encoding, code gen infra (#232)
* Combine ScionIPv4/v6 address types, binary encoding, code gen infra

Combine SCION IPv4/v6 addresses into one type. There is no reason to
separate these, as long as hybrid IPv4/6 ASes are not supported in
SCION. Even if such hybrid ASes would be supported, having separate
types still would seem to add little benefit.

Encode SCION addresses in binary for CBOR instead of as strings. It's
not even more complicated and the binary format is (apparently) just as
stable as the string representation.

Fix the SortedObjects code to actually return objects of all types and
remove the numObjects parameter -- passing in larger values would make
the object lists un-sorted because it's all based on strings that don't
sort nicely with more than 10 elements!
And avoid using IPv6 spelling of IPv4 addresses in the IPv6 test data --
that's just "funny", they were going to be represented as IPv4
addresses...

Additionally, sanitize the goyacc setup for the generation of
zoneFileDecoderGenerated.go based on zonefileParser.y I mean, there is
really no reason why this would have to muck around with sed.  Also, add
this to the _main_ Makefile, move the goyacc input file and make the
naming consistent.
Finally document the "go generate" tooling for the stringers  by adding
a step in the Makefile. Note: jsonenums turned out not to be work,
hope this is not going to be a problem in the future.
This is really just the minimum to keep my sanity the next time I'll
have to touch this.

* Sort IPs and SCION addresses numerically, fix testdata creation

Fix various issues in testdata creation (missing case for OTExtraKey) and the test code.
While messing with the testdata creation, we observed that IPs were not actually sorted correctly because they were sorted lexicographically, as strings.
Make the "Sort not implemented for external key"-case, whatever this is, from a log
error to a hard panic -- why _would_ it ever be sorted?